test: remove original testTensor
